// src/app/models/propertyModel.model.ts
export interface Property {
    code: number;
    type: string;
    address: string;
    neighborhood: string;
    city: string;
    state: string;
    area: number;
    dormitories: number;
    suites: number;
    garage: number;
    bathrooms: number;
    kitchens: number;
    leisureAreas: number;
    capacity: number;
    transactionType: string;
    iptu: number;
    rentPrice: number;
    salePrice: number;
    description: string;
    image: string;
}